+++ print_number_int_with_double_msg.h
@@ -2,7 +2,7 @@
 using namespace std;
 
 
-void PrintNumberIntWithDoubleMsg( string msgToPrintBefor,int numberFromFunc ,string msgToPrintAfter){
+void PrintNumberIntWithDoubleMsg( string msgToPrintBefor,float numberFromFunc ,string msgToPrintAfter){
 
     cout << msgToPrintBefor <<numberFromFunc  <<  msgToPrintAfter <<"\n";
